requirements.txt
google-cloud-bigquery
google-cloud-storage
functions-framework
pandas
db-dtypes
flask
